backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
SQLAlchemy>=2.0.32
asyncpg>=0.29.0
alembic>=1.13.2
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
pydantic>=2.7.0
pydantic-settings>=2.3.4
pandas>=2.2.2
APScheduler>=3.10.4

email-validator>=2.1.0

python-multipart>=0.0.9
